requirements.txt
absl-py==0.11.0
astunparse==1.6.3
bcrypt==3.2.0
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
colorama==0.4.4
cycler==0.10.0
ecdsa==0.14.1
fastapi==0.63.0
gast==0.3.3
google-auth==1.24.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
grpcio==1.35.0
h11==0.12.0
h5py==2.10.0
idna==2.10
importlib-metadata==3.4.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
loguru==0.5.3
Markdown==3.3.3
matplotlib==3.3.4
numpy==1.18.5
oauthlib==3.1.0
opt-einsum==3.3.0
passlib==1.7.4
Pillow==8.1.0
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydantic==1.7.3
pyparsing==2.4.7
python-dateutil==2.8.1
python-jose==3.2.0
python-multipart==0.0.5
PyYAML==5.3.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7
six==1.15.0
SQLAlchemy==1.3.23
starlette==0.13.6
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
tensorflow-cpu==2.3.1
tensorflow-estimator==2.3.0
termcolor==1.1.0
typing-extensions==3.7.4.3
urllib3==1.26.3
uvicorn==0.13.3
Werkzeug==1.0.1
win32-setctime==1.0.3
wrapt==1.12.1
zipp==3.4.0
